backend/requirements.txt
requests>=2.31.0
matplotlib>=3.8.0
networkx>=3.2.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
pyyaml>=6.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0
cachetools>=5.3.0
pydantic>=2.0.0
